Remove most uses of RangeProperties outside of engine_rocks (#8264)
* engine_traits: Stub out get_range_approximate_keys functions


* engine_traits: Derive Copy, Clone for Range


* engine_rocks: Implement get_range_approximate_keys


* raftstore: Use get_range_approximate_keys from RangePropertiesExt


* engine_traits: Stub out get_range_approximate_size


* engine_rocks: Implement get_range_approximate_size


* raftstore: Use get_range_approximate_size


* engine_traits: Stub out get_range_approximate_split_keys


* engine_rocks: Implement get_range_approximate_split_keys


* raftstore: Use get_range_approximate_split_keys


* engine_traits: Stub out get_range_approximate_middle


* engine_rocks: Implement get_range_approximate_middle


* raftstore: Use get_range_approximate_middle


* engine_traits: Stub out divide_range


* engine_rocks: Implement divide_range


* debug: Use divide_range from engine_traits


* rustfmt
