+++ components/engine_rocks/src/lib.rs
@@ -57,6 +57,8 @@ mod table_properties;
 pub use crate::table_properties::*;
 mod write_batch;
 pub use crate::write_batch::*;
+pub mod range_properties;
+pub use crate::range_properties::*;
 
 mod engine_iterator;
 pub use crate::engine_iterator::*;
