+++ components/raftstore/src/coprocessor/split_check/keys.rs
@@ -1,19 +1,16 @@
 // Copyright 2018 TiKV Project Authors. Licensed under Apache-2.0.
 
 use crate::store::{CasualMessage, CasualRouter};
-use engine_traits::CF_WRITE;
-use engine_traits::{KvEngine, Range, TableProperties, TablePropertiesCollection};
+use engine_traits::{KvEngine, Range};
 use kvproto::{metapb::Region, pdpb::CheckPolicy};
 use std::marker::PhantomData;
 use std::mem;
-use std::path::Path;
 use std::sync::{Arc, Mutex};
 
 use super::super::error::Result;
 use super::super::metrics::*;
 use super::super::{Coprocessor, KeyEntry, ObserverContext, SplitCheckObserver, SplitChecker};
 use super::Host;
-use engine_rocks::properties::{get_range_entries_and_versions, RangeProperties};
 
 pub struct Checker {
     max_keys_count: u64,
@@ -182,22 +179,14 @@ pub fn get_region_approximate_keys(
     region: &Region,
     large_threshold: u64,
 ) -> Result<u64> {
-    // try to get from RangeProperties first.
-    match get_region_approximate_keys_cf(db, CF_WRITE, region, large_threshold) {
-        Ok(v) => {
-            return Ok(v);
-        }
-        Err(e) => debug!(
-            "failed to get keys from RangeProperties";
-            "err" => ?e,
-        ),
-    }
-
     let start = keys::enc_start_key(region);
     let end = keys::enc_end_key(region);
-    let cf = box_try!(db.cf_handle(CF_WRITE));
-    let (_, keys) = get_range_entries_and_versions(db, cf, &start, &end).unwrap_or_default();
-    Ok(keys)
+    let range = Range::new(&start, &end);
+    Ok(box_try!(db.get_range_approximate_keys(
+        range,
+        region.get_id(),
+        large_threshold
+    )))
 }
 
 pub fn get_region_approximate_keys_cf(
@@ -206,46 +195,15 @@ pub fn get_region_approximate_keys_cf(
     region: &Region,
     large_threshold: u64,
 ) -> Result<u64> {
-    let start_key = keys::enc_start_key(region);
-    let end_key = keys::enc_end_key(region);
-    let range = Range::new(&start_key, &end_key);
-    let mut total_keys = 0;
-    let (mem_keys, _) = box_try!(db.get_approximate_memtable_stats_cf(cfname, &range));
-    total_keys += mem_keys;
-
-    let collection = box_try!(db.get_range_properties_cf(cfname, &start_key, &end_key));
-    for (_, v) in collection.iter() {
-        let props = box_try!(RangeProperties::decode(&v.user_collected_properties()));
-        total_keys += props.get_approximate_keys_in_range(&start_key, &end_key);
-    }
-
-    if large_threshold != 0 && total_keys > large_threshold {
-        let ssts = collection
-            .iter()
-            .map(|(k, v)| {
-                let props = RangeProperties::decode(&v.user_collected_properties()).unwrap();
-                let keys = props.get_approximate_keys_in_range(&start_key, &end_key);
-                format!(
-                    "{}:{}",
-                    Path::new(&*k)
-                        .file_name()
-                        .map(|f| f.to_str().unwrap())
-                        .unwrap_or(&*k),
-                    keys
-                )
-            })
-            .collect::<Vec<_>>()
-            .join(", ");
-        info!(
-            "region contains too many keys";
-            "region_id" => region.get_id(),
-            "total_keys" => total_keys,
-            "memtable" => mem_keys,
-            "ssts_keys" => ssts,
-            "cf" => cfname,
-        )
-    }
-    Ok(total_keys)
+    let start = keys::enc_start_key(region);
+    let end = keys::enc_end_key(region);
+    let range = Range::new(&start, &end);
+    Ok(box_try!(db.get_range_approximate_keys_cf(
+        cfname,
+        range,
+        region.get_id(),
+        large_threshold
+    )))
 }
 
 #[cfg(test)]
